# Worker: Babel ignore node_modules by default + some minor babel transform perf improvements