--- a/src/Worker.js
+++ b/src/Worker.js
@@ -52,6 +52,8 @@ function prepareJscodeshift(options) {
 }
 
 function setup(tr, babel) {
+  let babelRegister;
+
   if (babel === 'babel') {
     const presets = [];
     if (presetEnv) {
@@ -66,19 +68,24 @@ function setup(tr, babel) {
         require('@babel/preset-flow').default
     );
 
-    require('@babel/register')({
+    const plugins = [
+      require('@babel/plugin-proposal-class-properties').default,
+      require('@babel/plugin-proposal-nullish-coalescing-operator').default,
+      require('@babel/plugin-proposal-optional-chaining').default,
+      require('@babel/plugin-transform-modules-commonjs').default,
+    ];
+
+    babelRegister = require('@babel/register');
+
+    babelRegister({
       babelrc: false,
       presets,
-      plugins: [
-        require('@babel/plugin-proposal-class-properties').default,
-        require('@babel/plugin-proposal-nullish-coalescing-operator').default,
-        require('@babel/plugin-proposal-optional-chaining').default,
-        require('@babel/plugin-transform-modules-commonjs').default,
-      ],
+      plugins,
       extensions: [...DEFAULT_EXTENSIONS, '.ts', '.tsx'],
       // By default, babel register only compiles things inside the current working directory.
       // https://github.com/babel/babel/blob/2a4f16236656178e84b05b8915aab9261c55782c/packages/babel-register/src/node.js#L140-L157
       ignore: [
+        /\/node_modules\//,
         // Ignore parser related files
         /@babel\/parser/,
         /\/flow-parser\//,
@@ -89,6 +96,11 @@ function setup(tr, babel) {
   }
 
   const module = require(tr);
+
+  if (babelRegister) {
+    babelRegister.revert();
+  }
+
   transform = typeof module.default === 'function' ?
     module.default :
     module;